organization: move solver-only problem description to separate dir, move infrastructure (logging/profiling/etc) to separate dir
