--- src/common/common.hpp
@@ -1,23 +1,0 @@
-// Copyright (c) 2019-2020, Lawrence Livermore National Security, LLC and
-// other Serac Project Developers. See the top-level LICENSE file for
-// details.
-//
-// SPDX-License-Identifier: (BSD-3-Clause)
-
-/**
- * @file common.hpp
- *
- * @brief Aggregrates all the includes for "common" classes and functions
- */
-
-#ifndef COMMON
-#define COMMON
-
-#include "common/boundary_condition.hpp"
-#include "common/finite_element_state.hpp"
-#include "common/initialize.hpp"
-#include "common/logger.hpp"
-#include "common/serac_types.hpp"
-#include "common/terminator.hpp"
-
-#endif
