+++ src/drivers/serac.cpp
@@ -19,8 +19,10 @@
 #include "CLI11/CLI11.hpp"
 #include "coefficients/loading_functions.hpp"
 #include "coefficients/traction_coefficient.hpp"
-#include "common/common.hpp"
 #include "common/mesh_utils.hpp"
+#include "infrastructure/initialize.hpp"
+#include "infrastructure/logger.hpp"
+#include "infrastructure/terminator.hpp"
 #include "mfem.hpp"
 #include "serac_config.hpp"
 #include "solvers/nonlinear_solid_solver.hpp"
