+++ src/solvers/base_solver.cpp
@@ -10,8 +10,9 @@
 #include <fstream>
 #include <iostream>
 
-#include "common/common.hpp"
 #include "fmt/fmt.hpp"
+#include "infrastructure/logger.hpp"
+#include "infrastructure/terminator.hpp"
 
 namespace serac {
