+++ src/solvers/problem_description/equation_solver.hpp
