+++ src/solvers/problem_description/finite_element_state.cpp
@@ -4,7 +4,7 @@
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
 
-#include "common/finite_element_state.hpp"
+#include "solvers/problem_description/finite_element_state.hpp"
 
 namespace serac {
