+++ src/solvers/thermal_operators.hpp
@@ -15,9 +15,10 @@
 
 #include <memory>
 
-#include "common/common.hpp"
+#include "common/serac_types.hpp"
 #include "mfem.hpp"
-#include "solvers/equation_solver.hpp"
+#include "solvers/problem_description/boundary_condition.hpp"
+#include "solvers/problem_description/equation_solver.hpp"
 
 namespace serac {
